+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 2.15 branch
 ### 2.15.3
 * fixed code and a little bug for `ContactForm` and `ContactFormMailer` classes;
-* added tests for `Mailer` class.
+* added tests for `ContactFormMailer` and `Mailer` classes.
 
 ### 2.15.2
 * added `beforeMarshal()` method do `PostsTable` class. Removed